+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Accounting tool: long notes overwrite financial data.
 - Accounting tool: GMT, BST are not the same as Europe/London. ([#251](https://github.com/BittyTax/BittyTax/issues/251))
 - Binance parser: cash deposits should be gross amount. ([#257](https://github.com/BittyTax/BittyTax/issues/257))
+- Coinbase parser: handle commas in Buy Quantity matched from Note.
 ### Added
 - Conversion tool: generic parser added to support import from CryptoCredible tool.
 - Conversion tool: added parser for SnowTrace explorer.
@@ -50,6 +51,7 @@
 - Crypto.com parser: added new export format. ([#248](https://github.com/BittyTax/BittyTax/issues/248))
 - Nexo parser: added new export format. ([#246](https://github.com/BittyTax/BittyTax/issues/246))
 - Binance parser: added new deposits/withdrawals format. ([#258](https://github.com/BittyTax/BittyTax/issues/258))
+- Coinbase parser: new "Transaction history" data file format added.
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.